fix: move compute_distance to reports.py
